requirements.txt
# 核心框架
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
langgraph-cli>=0.1.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 可视化
matplotlib>=3.6.0
seaborn>=0.13.0

# 机器学习
lightgbm>=4.0.0
arch>=5.0.0
shap>=0.42.0

# 数据源
tushare>=1.4.0
requests>=2.28.0

# 阿里云 DashScope
dashscope>=1.20.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
cachetools>=5.3.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 代码质量（可选）
# flake8>=6.1.0
# mypy>=1.5.0
